Refactored code for conforming expressions to their default types
