+++ src/resolve/conform/mod.rs
@@ -4,6 +4,7 @@ mod from_integer;
 mod from_integer_literal;
 mod from_pointer;
 mod mode;
+pub mod to_default;
 
 pub use self::mode::ConformMode;
 use self::{
@@ -12,11 +13,10 @@ use self::{
 };
 use super::error::{ResolveError, ResolveErrorKind};
 use crate::{
-    ast::{ConformBehavior, FloatSize},
-    resolved::{Expr, ExprKind, Type, TypeKind, TypedExpr},
+    ast::ConformBehavior,
+    resolved::{Type, TypeKind, TypedExpr},
     source_files::Source,
 };
-use num_bigint::BigInt;
 
 pub fn conform_expr_or_error(
     expr: &TypedExpr,
@@ -60,50 +60,4 @@ pub fn conform_expr(
         TypeKind::Pointer(from_inner) => from_pointer(expr, mode, from_inner, to_type),
         _ => None,
     }
-}
-
-pub fn conform_expr_to_default(expr: TypedExpr) -> Result<TypedExpr, ResolveError> {
-    match &expr.resolved_type.kind {
-        TypeKind::IntegerLiteral(value) => {
-            conform_integer_literal_to_default_or_error(value, expr.expr.source)
-        }
-        TypeKind::FloatLiteral(value) => {
-            Ok(conform_float_literal_to_default(*value, expr.expr.source))
-        }
-        _ => Ok(expr),
-    }
-}
-
-fn conform_float_literal_to_default(value: f64, source: Source) -> TypedExpr {
-    TypedExpr::new(
-        TypeKind::f64().at(source),
-        Expr::new(ExprKind::FloatingLiteral(FloatSize::Bits64, value), source),
-    )
-}
-
-pub fn conform_integer_literal_to_default_or_error(
-    value: &BigInt,
-    source: Source,
-) -> Result<TypedExpr, ResolveError> {
-    integer_literal_to_default(value, source).ok_or_else(|| {
-        ResolveErrorKind::UnrepresentableInteger {
-            value: value.to_string(),
-        }
-        .at(source)
-    })
-}
-
-fn integer_literal_to_default(value: &BigInt, source: Source) -> Option<TypedExpr> {
-    for possible_type in [
-        TypeKind::i32().at(source),
-        TypeKind::u32().at(source),
-        TypeKind::i64().at(source),
-        TypeKind::u64().at(source),
-    ] {
-        if let Some(fit) = from_integer_literal(value, source, &possible_type) {
-            return Some(fit);
-        }
-    }
-
-    None
 }
